Promote the refresh endpoint fixture to the top level.
I'd like to test something with refreshed tokens outside of the views, and this seems like a reasonably simple way to get one.
