+++ tests/views/conftest.py
@@ -25,13 +25,3 @@ def call_auth_verify_endpoint(api_client, db):
         return api_client.post(path=url, data=data)
 
     return _call_auth_verify_endpoint
-
-
-@pytest.fixture
-def call_auth_refresh_endpoint(api_client, db):
-    def _call_auth_refresh_endpoint(token):
-        url = reverse("auth-refresh")
-        data = {"token": token}
-        return api_client.post(path=url, data=data)
-
-    return _call_auth_refresh_endpoint
